pyproject.toml
[project]
requires-python = ">=3.12"
name = "snaptop-mealprep"
version = "0.1.0"
description = "Agentic meal planning system using LangGraph, FastAPI, and BigQuery"
authors = ["leosneat@gmail.com"]
dependencies = [
	"fastapi>=0.109.0",
	"uvicorn[standard]>=0.27.0",
	"pydantic>=2.5.0",
	"google-cloud-bigquery>=3.20.0",
	"langchain_core",
	"langchain",
	"google-cloud-secret-manager>=2.25.0",
	"langchain-google-community>=3.0.0",
	"langchain-community>=0.4.1",
	"langchain-google-vertexai>=3.0.2",
	"google-generativeai>=0.8.5",
	"pytest>=9.0.0",
	"ruff>=0.14.4",
	"bs4>=0.0.2",
]

[tool.uv]

# pyproject.toml
[tool.pytest.ini_options]
pythonpath = [
	".",  # Add current directory
	"src", # Add a 'src' directory
]
